% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literals.R
\name{bdd_has_redundant_variables}
\alias{bdd_has_redundant_variables}
\title{Check expression for redundant variables}
\usage{
bdd_has_redundant_variables(expr)
}
\arguments{
\item{expr}{A character vector of length 1 containing a boolean expression in C++ syntax.}
}
\value{
Logical scalar indicating whether any variable is unnecessary.
}
\description{
Evaluates the given expression via `bdd_literals()` and returns `TRUE` when
at least one variable from the original expression is not required in the
reduced ROBDD (including the case where the expression is a constant).
}
\examples{
bdd_has_redundant_variables("x & !x")
bdd_has_redundant_variables("(y & x) | (y & !x)")

}
