% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_add.R
\name{cudd_add_epd_print_minterm}
\alias{cudd_add_epd_print_minterm}
\title{Print an EPD minterm count for an ADD}
\usage{
cudd_add_epd_print_minterm(add, nvars)
}
\arguments{
\item{add}{A [`CuddADD`] instance.}

\item{nvars}{Non-negative integer indicating how many variables to consider.}
}
\value{
`NULL`, invisibly.
}
\description{
This uses the CUDD `EpdPrintMinterm` implementation, which writes to R's
output stream.
}
