% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_add.R
\name{cudd_add_one}
\alias{cudd_add_one}
\title{Create an ADD node that represents logical TRUE}
\usage{
cudd_add_one(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
A [`CuddADD`] instance representing a constant TRUE ADD.
}
\description{
Create an ADD node that represents logical TRUE
}
