% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_conversions.R
\name{cudd_add_to_bdd}
\alias{cudd_add_to_bdd}
\title{Convert an ADD to a BDD using the non-zero pattern}
\usage{
cudd_add_to_bdd(add)
}
\arguments{
\item{add}{A [`CuddADD`] instance.}
}
\value{
A [`CuddBDD`] instance.
}
\description{
Convert an ADD to a BDD using the non-zero pattern
}
