% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_add.R
\name{cudd_add_var}
\alias{cudd_add_var}
\title{Create or access an ADD variable}
\usage{
cudd_add_var(manager, index = NULL)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}

\item{index}{Optional non-negative integer index of the ADD variable.}
}
\value{
A [`CuddADD`] instance representing the requested variable.
}
\description{
When `index` is `NULL`, a new ADD variable is created in the manager and
returned. Otherwise, the ADD variable at the specified index is returned.
}
