% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_add.R
\name{cudd_add_zero}
\alias{cudd_add_zero}
\title{Create an ADD node that represents logical FALSE}
\usage{
cudd_add_zero(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
A [`CuddADD`] instance representing a constant FALSE ADD.
}
\description{
Create an ADD node that represents logical FALSE
}
