% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_bdd.R
\name{cudd_bdd_epd_print_minterm}
\alias{cudd_bdd_epd_print_minterm}
\title{Print an EPD minterm count for a BDD}
\usage{
cudd_bdd_epd_print_minterm(bdd, nvars)
}
\arguments{
\item{bdd}{A [`CuddBDD`] instance.}

\item{nvars}{Non-negative integer indicating how many variables to consider.}
}
\value{
`NULL`, invisibly.
}
\description{
This uses the CUDD `EpdPrintMinterm` implementation, which writes to R's
output stream.
}
