% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_bdd_methods.R
\name{cudd_bdd_methods}
\alias{cudd_bdd_methods}
\alias{cudd_bdd_is_zero}
\alias{cudd_bdd_is_var}
\alias{cudd_bdd_leq}
\alias{cudd_bdd_and_abstract}
\alias{cudd_bdd_exist_abstract}
\alias{cudd_bdd_univ_abstract}
\alias{cudd_bdd_xor_exist_abstract}
\alias{cudd_bdd_boolean_diff}
\alias{cudd_bdd_var_is_dependent}
\alias{cudd_bdd_ite}
\alias{cudd_bdd_ite_constant}
\alias{cudd_bdd_intersect}
\alias{cudd_bdd_and_limit}
\alias{cudd_bdd_or_limit}
\alias{cudd_bdd_nand}
\alias{cudd_bdd_nor}
\alias{cudd_bdd_xnor}
\alias{cudd_bdd_cofactor}
\alias{cudd_bdd_constrain}
\alias{cudd_bdd_compose}
\alias{cudd_bdd_permute}
\alias{cudd_bdd_swap_variables}
\alias{cudd_bdd_vector_compose}
\alias{cudd_bdd_li_compaction}
\alias{cudd_bdd_squeeze}
\alias{cudd_bdd_interpolate}
\alias{cudd_bdd_minimize}
\alias{cudd_bdd_subset_compress}
\alias{cudd_bdd_superset_compress}
\alias{cudd_bdd_literal_set_intersection}
\alias{cudd_bdd_c_projection}
\alias{cudd_bdd_min_hamming_dist}
\alias{cudd_bdd_eval}
\alias{cudd_bdd_decreasing}
\alias{cudd_bdd_increasing}
\alias{cudd_bdd_make_prime}
\alias{cudd_bdd_subset_heavy_branch}
\alias{cudd_bdd_superset_heavy_branch}
\alias{cudd_bdd_subset_short_paths}
\alias{cudd_bdd_superset_short_paths}
\alias{cudd_bdd_print_cover}
\alias{cudd_bdd_print_cover_with_cube}
\alias{cudd_bdd_pick_one_cube}
\alias{cudd_bdd_pick_one_minterm}
\alias{cudd_bdd_isop}
\alias{cudd_bdd_port_to_zdd}
\alias{cudd_bdd_factored_form_string}
\alias{cudd_bdd_print_factored_form}
\title{Additional BDD methods}
\usage{
cudd_bdd_is_zero(bdd)

cudd_bdd_is_var(bdd)

cudd_bdd_leq(bdd, other)

cudd_bdd_and_abstract(bdd, other, cube, limit = 0L)

cudd_bdd_exist_abstract(bdd, cube, limit = 0L)

cudd_bdd_univ_abstract(bdd, cube)

cudd_bdd_xor_exist_abstract(bdd, other, cube)

cudd_bdd_boolean_diff(bdd, index)

cudd_bdd_var_is_dependent(bdd, var)

cudd_bdd_ite(bdd, g, h, limit = 0L)

cudd_bdd_ite_constant(bdd, g, h)

cudd_bdd_intersect(bdd, other)

cudd_bdd_and_limit(bdd, other, limit = 0L)

cudd_bdd_or_limit(bdd, other, limit = 0L)

cudd_bdd_nand(bdd, other)

cudd_bdd_nor(bdd, other)

cudd_bdd_xnor(bdd, other, limit = 0L)

cudd_bdd_cofactor(bdd, other)

cudd_bdd_constrain(bdd, other)

cudd_bdd_compose(bdd, other, index)

cudd_bdd_permute(bdd, permut)

cudd_bdd_swap_variables(bdd, x, y)

cudd_bdd_vector_compose(bdd, vector)

cudd_bdd_li_compaction(bdd, other)

cudd_bdd_squeeze(bdd, other)

cudd_bdd_interpolate(bdd, other)

cudd_bdd_minimize(bdd, other)

cudd_bdd_subset_compress(bdd, nvars, threshold)

cudd_bdd_superset_compress(bdd, nvars, threshold)

cudd_bdd_literal_set_intersection(bdd, other)

cudd_bdd_c_projection(bdd, other)

cudd_bdd_min_hamming_dist(bdd, minterm, upper_bound)

cudd_bdd_eval(bdd, inputs)

cudd_bdd_decreasing(bdd, index)

cudd_bdd_increasing(bdd, index)

cudd_bdd_make_prime(bdd, other)

cudd_bdd_subset_heavy_branch(bdd, num_vars, threshold)

cudd_bdd_superset_heavy_branch(bdd, num_vars, threshold)

cudd_bdd_subset_short_paths(bdd, num_vars, threshold, hardlimit = FALSE)

cudd_bdd_superset_short_paths(bdd, num_vars, threshold, hardlimit = FALSE)

cudd_bdd_print_cover(bdd)

cudd_bdd_print_cover_with_cube(bdd, cube)

cudd_bdd_pick_one_cube(bdd)

cudd_bdd_pick_one_minterm(bdd, vars)

cudd_bdd_isop(bdd, upper)

cudd_bdd_port_to_zdd(bdd)

cudd_bdd_factored_form_string(bdd)

cudd_bdd_print_factored_form(bdd)
}
\arguments{
\item{bdd}{A [`CuddBDD`] instance.}

\item{other, cube, var, g, h, upper}{Additional [`CuddBDD`] instances used by the
operation.}

\item{limit}{Optional non-negative integer limit passed to CUDD.}

\item{index}{Integer index values used by the operation.}

\item{permut}{Integer vector describing a variable permutation.}

\item{x, y, vector, vars}{Lists of [`CuddBDD`] instances.}

\item{nvars, num_vars, threshold}{Non-negative integer values used by the operation.}

\item{minterm, inputs}{Integer vectors used by the operation.}

\item{upper_bound}{Integer upper bound for min hamming distance.}

\item{hardlimit}{Logical scalar for short path routines.}
}
\value{
A [`CuddBDD`] instance, a logical scalar, numeric scalar, character
  scalar, or list depending on the operation.
}
\description{
Convenience wrappers for additional methods on CUDD BDD objects.
}
