% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_bdd.R
\name{cudd_bdd_one}
\alias{cudd_bdd_one}
\title{Create a BDD node that represents logical TRUE}
\usage{
cudd_bdd_one(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
A [`CuddBDD`] instance representing a constant TRUE BDD.
}
\description{
Create a BDD node that represents logical TRUE
}
