% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_bdd.R
\name{cudd_bdd_print_debug}
\alias{cudd_bdd_print_debug}
\title{Print a debug representation for a BDD}
\usage{
cudd_bdd_print_debug(bdd, nvars = NULL, verbosity = NULL)
}
\arguments{
\item{bdd}{A [`CuddBDD`] instance.}

\item{nvars}{Optional non-negative integer indicating how many variables to
include. Defaults to the manager size.}

\item{verbosity}{Optional non-negative integer debug verbosity.}
}
\value{
`NULL`, invisibly.
}
\description{
This uses the CUDD `PrintDebug` implementation, which writes to R's output
stream.
}
