% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_bdd.R
\name{cudd_bdd_print_minterm}
\alias{cudd_bdd_print_minterm}
\title{Print the minterm representation for a BDD}
\usage{
cudd_bdd_print_minterm(bdd)
}
\arguments{
\item{bdd}{A [`CuddBDD`] instance.}
}
\value{
`NULL`, invisibly.
}
\description{
This uses the CUDD `PrintMinterm` implementation, which writes to R's
output stream.
}
