% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_bdd.R
\name{cudd_bdd_restrict}
\alias{cudd_bdd_restrict}
\title{Restrict a BDD with a constraint}
\usage{
cudd_bdd_restrict(bdd, constraint)
}
\arguments{
\item{bdd}{A [`CuddBDD`] instance to restrict.}

\item{constraint}{A [`CuddBDD`] instance expressing the restriction.}
}
\value{
A [`CuddBDD`] instance.
}
\description{
Applies the CUDD `Restrict` operator to reduce a BDD given a constraint BDD.
}
