% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_conversions.R
\name{cudd_bdd_to_add}
\alias{cudd_bdd_to_add}
\title{Convert a BDD to an ADD}
\usage{
cudd_bdd_to_add(bdd)
}
\arguments{
\item{bdd}{A [`CuddBDD`] instance.}
}
\value{
A [`CuddADD`] instance.
}
\description{
Convert a BDD to an ADD
}
