% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_conversions.R
\name{cudd_bdd_to_zdd}
\alias{cudd_bdd_to_zdd}
\title{Convert a BDD to a ZDD}
\usage{
cudd_bdd_to_zdd(bdd)
}
\arguments{
\item{bdd}{A [`CuddBDD`] instance.}
}
\value{
A [`CuddZDD`] instance.
}
\description{
Convert a BDD to a ZDD
}
