% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_bdd.R
\name{cudd_bdd_var}
\alias{cudd_bdd_var}
\title{Create or access a BDD variable}
\usage{
cudd_bdd_var(manager, index = NULL)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}

\item{index}{Optional non-negative integer index of the BDD variable.}
}
\value{
A [`CuddBDD`] instance representing the requested variable.
}
\description{
When `index` is `NULL`, a new BDD variable is created in the manager and
returned. Otherwise, the BDD variable at the specified index is returned.
}
