% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_bdd.R
\name{cudd_bdd_zero}
\alias{cudd_bdd_zero}
\title{Create a BDD node that represents logical FALSE}
\usage{
cudd_bdd_zero(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
A [`CuddBDD`] instance representing a constant FALSE BDD.
}
\description{
Create a BDD node that represents logical FALSE
}
