% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_manager.R
\name{cudd_set_min_hit}
\alias{cudd_set_min_hit}
\alias{cudd_set_loose_up_to}
\alias{cudd_make_verbose}
\alias{cudd_make_terse}
\alias{cudd_is_verbose}
\alias{cudd_info}
\alias{cudd_push_variable_name}
\alias{cudd_clear_variable_names}
\alias{cudd_get_variable_name}
\alias{cudd_manager_controls}
\alias{cudd_read_start_time}
\alias{cudd_read_elapsed_time}
\alias{cudd_set_start_time}
\alias{cudd_reset_start_time}
\alias{cudd_read_time_limit}
\alias{cudd_set_time_limit}
\alias{cudd_update_time_limit}
\alias{cudd_increase_time_limit}
\alias{cudd_unset_time_limit}
\alias{cudd_time_limited}
\alias{cudd_autodyn_enable}
\alias{cudd_autodyn_disable}
\alias{cudd_reordering_status}
\alias{cudd_autodyn_enable_zdd}
\alias{cudd_autodyn_disable_zdd}
\alias{cudd_reordering_status_zdd}
\alias{cudd_reduce_heap}
\alias{cudd_zdd_realignment_enabled}
\alias{cudd_zdd_realign_enable}
\alias{cudd_zdd_realign_disable}
\alias{cudd_bdd_realignment_enabled}
\alias{cudd_bdd_realign_enable}
\alias{cudd_bdd_realign_disable}
\alias{cudd_background}
\alias{cudd_set_background}
\alias{cudd_read_max_cache}
\alias{cudd_read_max_cache_hard}
\alias{cudd_set_max_cache_hard}
\alias{cudd_read_slots}
\alias{cudd_read_keys}
\alias{cudd_read_dead}
\alias{cudd_read_min_dead}
\alias{cudd_read_max_reorderings}
\alias{cudd_set_max_reorderings}
\alias{cudd_read_reordering_time}
\alias{cudd_read_garbage_collections}
\alias{cudd_read_garbage_collection_time}
\alias{cudd_read_sift_max_var}
\alias{cudd_set_sift_max_var}
\alias{cudd_read_sift_max_swap}
\alias{cudd_set_sift_max_swap}
\alias{cudd_read_max_growth}
\alias{cudd_set_max_growth}
\alias{cudd_read_perm}
\alias{cudd_read_perm_zdd}
\alias{cudd_read_inv_perm}
\alias{cudd_read_inv_perm_zdd}
\alias{cudd_read_vars}
\alias{cudd_read_epsilon}
\alias{cudd_set_epsilon}
\alias{cudd_read_groupcheck}
\alias{cudd_set_groupcheck}
\alias{cudd_garbage_collection_enabled}
\alias{cudd_enable_garbage_collection}
\alias{cudd_disable_garbage_collection}
\alias{cudd_dead_are_counted}
\alias{cudd_turn_on_count_dead}
\alias{cudd_turn_off_count_dead}
\alias{cudd_read_recomb}
\alias{cudd_set_recomb}
\alias{cudd_read_symmviolation}
\alias{cudd_set_symmviolation}
\alias{cudd_read_arcviolation}
\alias{cudd_set_arcviolation}
\alias{cudd_read_population_size}
\alias{cudd_set_population_size}
\alias{cudd_read_number_xovers}
\alias{cudd_set_number_xovers}
\alias{cudd_read_order_randomization}
\alias{cudd_set_order_randomization}
\alias{cudd_read_memory_in_use}
\alias{cudd_read_peak_node_count}
\alias{cudd_read_node_count_current}
\alias{cudd_read_node_count_zdd}
\alias{cudd_enable_reordering_reporting}
\alias{cudd_disable_reordering_reporting}
\alias{cudd_reordering_reporting}
\alias{cudd_read_error_code}
\alias{cudd_clear_error_code}
\alias{cudd_read_next_reordering}
\alias{cudd_set_next_reordering}
\alias{cudd_read_swap_steps}
\alias{cudd_read_max_live}
\alias{cudd_set_max_live}
\alias{cudd_read_max_memory}
\alias{cudd_set_max_memory}
\alias{cudd_bdd_bind_var}
\alias{cudd_bdd_unbind_var}
\alias{cudd_bdd_var_is_bound}
\title{CUDD manager controls}
\usage{
cudd_set_min_hit(manager, hr)

cudd_set_loose_up_to(manager, lut)

cudd_make_verbose(manager)

cudd_make_terse(manager)

cudd_is_verbose(manager)

cudd_info(manager)

cudd_push_variable_name(manager, name)

cudd_clear_variable_names(manager)

cudd_get_variable_name(manager, index)

cudd_read_start_time(manager)

cudd_read_elapsed_time(manager)

cudd_set_start_time(manager, st)

cudd_reset_start_time(manager)

cudd_read_time_limit(manager)

cudd_set_time_limit(manager, tl)

cudd_update_time_limit(manager)

cudd_increase_time_limit(manager, inc)

cudd_unset_time_limit(manager)

cudd_time_limited(manager)

cudd_autodyn_enable(manager, method = NULL)

cudd_autodyn_disable(manager)

cudd_reordering_status(manager)

cudd_autodyn_enable_zdd(manager, method = NULL)

cudd_autodyn_disable_zdd(manager)

cudd_reordering_status_zdd(manager)

cudd_reduce_heap(manager, heuristic = NULL, minsize = 0L)

cudd_zdd_realignment_enabled(manager)

cudd_zdd_realign_enable(manager)

cudd_zdd_realign_disable(manager)

cudd_bdd_realignment_enabled(manager)

cudd_bdd_realign_enable(manager)

cudd_bdd_realign_disable(manager)

cudd_background(manager)

cudd_set_background(manager, add)

cudd_read_max_cache(manager)

cudd_read_max_cache_hard(manager)

cudd_set_max_cache_hard(manager, mc)

cudd_read_slots(manager)

cudd_read_keys(manager)

cudd_read_dead(manager)

cudd_read_min_dead(manager)

cudd_read_max_reorderings(manager)

cudd_set_max_reorderings(manager, mr)

cudd_read_reordering_time(manager)

cudd_read_garbage_collections(manager)

cudd_read_garbage_collection_time(manager)

cudd_read_sift_max_var(manager)

cudd_set_sift_max_var(manager, smv)

cudd_read_sift_max_swap(manager)

cudd_set_sift_max_swap(manager, sms)

cudd_read_max_growth(manager)

cudd_set_max_growth(manager, mg)

cudd_read_perm(manager, i)

cudd_read_perm_zdd(manager, i)

cudd_read_inv_perm(manager, i)

cudd_read_inv_perm_zdd(manager, i)

cudd_read_vars(manager, i)

cudd_read_epsilon(manager)

cudd_set_epsilon(manager, ep)

cudd_read_groupcheck(manager)

cudd_set_groupcheck(manager, gc)

cudd_garbage_collection_enabled(manager)

cudd_enable_garbage_collection(manager)

cudd_disable_garbage_collection(manager)

cudd_dead_are_counted(manager)

cudd_turn_on_count_dead(manager)

cudd_turn_off_count_dead(manager)

cudd_read_recomb(manager)

cudd_set_recomb(manager, recomb)

cudd_read_symmviolation(manager)

cudd_set_symmviolation(manager, symm)

cudd_read_arcviolation(manager)

cudd_set_arcviolation(manager, arc)

cudd_read_population_size(manager)

cudd_set_population_size(manager, pop)

cudd_read_number_xovers(manager)

cudd_set_number_xovers(manager, xovers)

cudd_read_order_randomization(manager)

cudd_set_order_randomization(manager, factor)

cudd_read_memory_in_use(manager)

cudd_read_peak_node_count(manager)

cudd_read_node_count_current(manager)

cudd_read_node_count_zdd(manager)

cudd_enable_reordering_reporting(manager)

cudd_disable_reordering_reporting(manager)

cudd_reordering_reporting(manager)

cudd_read_error_code(manager)

cudd_clear_error_code(manager)

cudd_read_next_reordering(manager)

cudd_set_next_reordering(manager, nr)

cudd_read_swap_steps(manager)

cudd_read_max_live(manager)

cudd_set_max_live(manager, max_live)

cudd_read_max_memory(manager)

cudd_set_max_memory(manager, max_mem)

cudd_bdd_bind_var(manager, index)

cudd_bdd_unbind_var(manager, index)

cudd_bdd_var_is_bound(manager, index)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}

\item{hr, lut}{Integers used for cache hit/loose-up-to thresholds.}

\item{name}{Single string for variable names.}

\item{st, tl, inc}{Numeric scalars used for time limit operations.}

\item{method}{Optional integer reordering method (CUDD enum value).}

\item{heuristic}{Integer reordering heuristic (CUDD enum value).}

\item{minsize}{Integer minimum size for reordering.}

\item{add}{A [`CuddADD`] instance.}

\item{mc, mr, smv, sms, max_live}{Integers used for cache/reordering limits.}

\item{mg, ep}{Numeric values for growth/epsilon settings.}

\item{i, index}{Integer indices for variable-related queries.}

\item{gc}{Integer CUDD aggregation type.}

\item{recomb, symm, arc, pop, xovers}{Integers for genetic algorithm settings.}

\item{factor}{Integer randomization factor.}

\item{nr}{Integer next reordering value.}

\item{max_mem}{Numeric memory limit.}
}
\value{
A scalar value or `NULL`, depending on the call.
}
\description{
Convenience wrappers around the CUDD manager API for tuning reordering,
time limits, memory limits, and reading manager statistics.
}
\details{
Reordering heuristic integers follow the CUDD enum ordering:
`0` = `CUDD_REORDER_SAME`, `1` = `CUDD_REORDER_NONE`, `2` = `CUDD_REORDER_RANDOM`,
`3` = `CUDD_REORDER_RANDOM_PIVOT`, `4` = `CUDD_REORDER_SIFT`,
`5` = `CUDD_REORDER_SIFT_CONVERGE`, `6` = `CUDD_REORDER_SYMM_SIFT`,
`7` = `CUDD_REORDER_SYMM_SIFT_CONV`, `8` = `CUDD_REORDER_WINDOW2`,
`9` = `CUDD_REORDER_WINDOW3`, `10` = `CUDD_REORDER_WINDOW4`,
`11` = `CUDD_REORDER_WINDOW2_CONV`, `12` = `CUDD_REORDER_WINDOW3_CONV`,
`13` = `CUDD_REORDER_WINDOW4_CONV`, `14` = `CUDD_REORDER_GROUP_SIFT`,
`15` = `CUDD_REORDER_GROUP_SIFT_CONV`, `16` = `CUDD_REORDER_ANNEALING`,
`17` = `CUDD_REORDER_GENETIC`, `18` = `CUDD_REORDER_LINEAR`,
`19` = `CUDD_REORDER_LINEAR_CONVERGE`, `20` = `CUDD_REORDER_LAZY_SIFT`,
`21` = `CUDD_REORDER_EXACT`.
}
