% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_manager.R
\name{cudd_read_cache_hits}
\alias{cudd_read_cache_hits}
\title{Read the cache hit count}
\usage{
cudd_read_cache_hits(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
Double scalar with the cache hit count.
}
\description{
Read the cache hit count
}
