% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_manager.R
\name{cudd_read_cache_lookups}
\alias{cudd_read_cache_lookups}
\title{Read the cache lookup count}
\usage{
cudd_read_cache_lookups(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
Double scalar with the cache lookup count.
}
\description{
Read the cache lookup count
}
