% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_manager.R
\name{cudd_read_cache_used_slots}
\alias{cudd_read_cache_used_slots}
\title{Read the cache used slots}
\usage{
cudd_read_cache_used_slots(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
Double scalar with the cache used slot count.
}
\description{
Read the cache used slots
}
