% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_manager.R
\name{cudd_read_loose_up_to}
\alias{cudd_read_loose_up_to}
\title{Read the cache loose up-to threshold}
\usage{
cudd_read_loose_up_to(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
Integer scalar with the cache loose up-to threshold.
}
\description{
Read the cache loose up-to threshold
}
