% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_manager.R
\name{cudd_read_min_hit}
\alias{cudd_read_min_hit}
\title{Read the minimum cache hit ratio}
\usage{
cudd_read_min_hit(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
Integer scalar with the minimum cache hit ratio.
}
\description{
Read the minimum cache hit ratio
}
