% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_manager.R
\name{cudd_read_node_count}
\alias{cudd_read_node_count}
\title{Read the live node count}
\usage{
cudd_read_node_count(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
Double scalar with the live node count.
}
\description{
Read the live node count
}
