% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_manager.R
\name{cudd_read_reorderings}
\alias{cudd_read_reorderings}
\title{Read the number of reorderings performed}
\usage{
cudd_read_reorderings(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
Integer scalar with the number of reorderings.
}
\description{
Read the number of reorderings performed
}
