% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_manager.R
\name{cudd_read_zdd_size}
\alias{cudd_read_zdd_size}
\title{Read the number of ZDD variables in the manager}
\usage{
cudd_read_zdd_size(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
Integer scalar with the number of ZDD variables.
}
\description{
Read the number of ZDD variables in the manager
}
