% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_zdd.R
\name{cudd_zdd_one}
\alias{cudd_zdd_one}
\title{Create a ZDD node that represents the constant one}
\usage{
cudd_zdd_one(manager, index = 0L)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}

\item{index}{Non-negative integer index for the ZDD one node.}
}
\value{
A [`CuddZDD`] instance representing the ZDD one node.
}
\description{
Create a ZDD node that represents the constant one
}
