% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_zdd.R
\name{cudd_zdd_print_minterm}
\alias{cudd_zdd_print_minterm}
\title{Print the minterm representation for a ZDD}
\usage{
cudd_zdd_print_minterm(zdd)
}
\arguments{
\item{zdd}{A [`CuddZDD`] instance.}
}
\value{
`NULL`, invisibly.
}
\description{
This uses the CUDD `PrintMinterm` implementation, which writes to R's
output stream.
}
