% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_conversions.R
\name{cudd_zdd_to_bdd}
\alias{cudd_zdd_to_bdd}
\title{Convert a ZDD to a BDD}
\usage{
cudd_zdd_to_bdd(zdd)
}
\arguments{
\item{zdd}{A [`CuddZDD`] instance.}
}
\value{
A [`CuddBDD`] instance.
}
\description{
Convert a ZDD to a BDD
}
