% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_zdd.R
\name{cudd_zdd_zero}
\alias{cudd_zdd_zero}
\title{Create a ZDD node that represents the constant zero}
\usage{
cudd_zdd_zero(manager)
}
\arguments{
\item{manager}{A [`CuddManager`] instance.}
}
\value{
A [`CuddZDD`] instance representing the ZDD zero node.
}
\description{
Create a ZDD node that represents the constant zero
}
