% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_bdd.R
\docType{class}
\name{CuddBDD-class}
\alias{CuddBDD-class}
\alias{show,CuddBDD-method}
\title{S4 wrapper for a CUDD BDD}
\usage{
\S4method{show}{CuddBDD}(object)
}
\arguments{
\item{object}{A `CuddBDD` instance.}
}
\description{
The BDD wrapper encapsulates the underlying C++ `BDD` instance and is
represented in R as an external pointer managed by a finalizer.
}
\section{Functions}{
\itemize{
\item \code{show(CuddBDD)}: Show a brief summary of the BDD.

}}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to the underlying BDD object.}
}}

\keyword{internal}
