% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_manager.R
\docType{class}
\name{CuddManager-class}
\alias{CuddManager-class}
\alias{show,CuddManager-method}
\title{S4 wrapper for a CUDD manager}
\usage{
\S4method{show}{CuddManager}(object)
}
\arguments{
\item{object}{A `CuddManager` instance.}
}
\description{
The CUDD manager encapsulates the underlying C++ `Cudd` instance and is
represented in R as an external pointer managed by a finalizer.
}
\section{Functions}{
\itemize{
\item \code{show(CuddManager)}: Show a brief summary of the manager.

}}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to the underlying CUDD manager.}
}}

\keyword{internal}
