% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_zdd.R
\docType{class}
\name{CuddZDD-class}
\alias{CuddZDD-class}
\alias{show,CuddZDD-method}
\alias{*,CuddZDD,CuddZDD-method}
\alias{+,CuddZDD,CuddZDD-method}
\alias{^,CuddZDD,CuddZDD-method}
\alias{!,CuddZDD-method}
\title{S4 wrapper for a CUDD ZDD}
\usage{
\S4method{show}{CuddZDD}(object)

\S4method{*}{CuddZDD,CuddZDD}(e1, e2)

\S4method{+}{CuddZDD,CuddZDD}(e1, e2)

\S4method{^}{CuddZDD,CuddZDD}(e1, e2)

\S4method{!}{CuddZDD}(x)
}
\arguments{
\item{object}{A `CuddZDD` instance.}

\item{e1}{A `CuddZDD` instance.}

\item{e2}{A `CuddZDD` instance.}

\item{x}{A `CuddZDD` instance.}
}
\value{
A `CuddZDD` instance.

A `CuddZDD` instance.

An error indicating the operator is unsupported.

An error indicating the operator is unsupported.
}
\description{
The ZDD wrapper encapsulates the underlying C++ `ZDD` instance and is
represented in R as an external pointer managed by a finalizer.
}
\section{Functions}{
\itemize{
\item \code{show(CuddZDD)}: Show a brief summary of the ZDD.

\item \code{e1 * e2}: Combine ZDDs with intersection

\item \code{e1 + e2}: Combine ZDDs with union

\item \code{e1^e2}: XOR is not defined for ZDDs

\item \code{`!`(CuddZDD)}: Negation is not defined for ZDDs

}}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to the underlying ZDD object.}

\item{\code{manager_ptr}}{External pointer to the owning CUDD manager.}
}}

\keyword{internal}
