% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_add.R
\name{print,CuddADD-method}
\alias{print,CuddADD-method}
\title{Print the minterm representation for an ADD}
\usage{
\S4method{print}{CuddADD}(x, ...)
}
\arguments{
\item{x}{A [`CuddADD`] instance.}

\item{...}{Unused.}
}
\value{
The input `x`, invisibly.
}
\description{
This uses the CUDD `PrintMinterm` implementation, which writes to R's
output stream.
}
