% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_bdd.R
\name{print,CuddBDD-method}
\alias{print,CuddBDD-method}
\title{Print the minterm representation for a BDD}
\usage{
\S4method{print}{CuddBDD}(x, ...)
}
\arguments{
\item{x}{A [`CuddBDD`] instance.}

\item{...}{Unused.}
}
\value{
The input `x`, invisibly.
}
\description{
This uses the CUDD `PrintMinterm` implementation, which writes to R's
output stream.
}
