% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_zdd.R
\name{print,CuddZDD-method}
\alias{print,CuddZDD-method}
\title{Print the minterm representation for a ZDD}
\usage{
\S4method{print}{CuddZDD}(x, ...)
}
\arguments{
\item{x}{A [`CuddZDD`] instance.}

\item{...}{Unused.}
}
\value{
The input `x`, invisibly.
}
\description{
This uses the CUDD `PrintMinterm` implementation, which writes to R's
output stream.
}
