% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudd_add.R
\docType{class}
\name{CuddADD-class}
\alias{CuddADD-class}
\alias{show,CuddADD-method}
\title{S4 wrapper for a CUDD ADD}
\usage{
\S4method{show}{CuddADD}(object)
}
\arguments{
\item{object}{A `CuddADD` instance.}
}
\description{
The ADD wrapper encapsulates the underlying C++ `ADD` instance and is
represented in R as an external pointer managed by a finalizer.
}
\section{Functions}{
\itemize{
\item \code{show(CuddADD)}: Show a brief summary of the ADD.

}}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to the underlying ADD object.}
}}

\keyword{internal}
